requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.4.0
Authlib==1.6.5
backoff==2.2.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
coverage==7.10.6
cryptography==46.0.3
cycler==0.12.1
cyclopts==3.24.0
distro==1.9.0
dnspython==2.8.0
docker==7.1.0
docstring_parser==0.17.0
docutils==0.22.2
email-validator==2.3.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastmcp==2.12.5
fonttools==4.60.1
googleapis-common-protos==1.72.0
graphviz==0.21
greenlet==3.2.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
influxdb3-python==0.16.0

iniconfig==2.1.0
isodate==0.7.2
jiter==0.11.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langchain==0.3.27
langchain-core==0.3.76
langchain-openai==0.3.33
langchain-text-splitters==0.3.11
langfuse==2.37.0
langgraph==0.6.8
langgraph-checkpoint==2.1.1
langgraph-checkpoint-sqlite==2.0.11
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langsmith==0.4.28
lazy-object-proxy==1.12.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.6
mcp==1.16.0
mdurl==0.1.2
more-itertools==10.8.0
numpy==2.3.3
openai==1.108.0
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
openpyxl==3.1.5
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.3
ormsgpack==1.10.0
packaging==25.0
pandas==2.3.2
parse==1.20.2
pathable==0.4.4
pillow==11.3.0
pluggy==1.6.0
protobuf==6.33.0
psycopg2-binary==2.9.10
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.5
pyperclip==1.11.0
pytest==8.4.2
pytest-cov==7.0.0
pytest-mock==3.15.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pywin32==311
PyYAML==6.0.2
reactivex==4.0.4
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
responses==0.25.8
rfc3339-validator==0.1.4
rich==14.2.0
rich-rst==1.3.2
rpds-py==0.27.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlite-vec==0.1.6
sse-starlette==3.0.2
starlette==0.48.0
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
Werkzeug==3.1.1

wrapt==1.17.3
xxhash==3.5.0
zipp==3.23.0
zstandard==0.25.0

# Plan A Automation Dependencies
watchdog==6.0.0
schedule==1.2.2

# OAuth2 Email Support (Microsoft Graph API)
requests==2.32.5

# Web API Upload Service
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
